requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
djangorestframework
python-dotenv
django-cors-headers